+++ seatable-app-old/src/triggers/row_update.js
@@ -1,5 +1,5 @@
-const ctx = require("../ctx");
-const _ = require("lodash");
+const ctx = require('../ctx');
+const _ = require('lodash');
 
 /**
  * perform
@@ -91,7 +91,7 @@ const perform = async (z, bundle) => {
    * */
   const response = await z.request({
     url: `${bundle.authData.server}/dtable-server/api/v1/dtables/${dtableCtx.dtable_uuid}/rows/`,
-    headers: {Authorization: `Token ${dtableCtx.access_token}`},
+    headers: { Authorization: `Token ${dtableCtx.access_token}` },
     params: ctx.requestParamsBundle(bundle),
   });
 
@@ -102,26 +102,33 @@ const perform = async (z, bundle) => {
   }
 
   // usually the columns have the name _ctime and _mtime. But if these columns are set, they have different names.
-  let mtime_column_name = "_mtime"; 
-  const mtime = _.find(tableMetadata.columns, ["type", "mtime"]);
+  let mtime_column_name = '_mtime';
+  const mtime = _.find(tableMetadata.columns, ['type', 'mtime']);
   if (undefined !== mtime) {
-    mtime_column_name = mtime.name; 
+    mtime_column_name = mtime.name;
   }
 
   // limit payload size
   // https://platform.zapier.com/docs/constraints#payload-size-triggers
   // rows.reverse();
-  rows = _.orderBy(rows, [mtime_column_name], ["desc"]);
+  rows = _.orderBy(rows, [mtime_column_name], ['desc']);
   if (bundle.meta && bundle.meta.isLoadingSample) {
     rows.splice(bundle.meta.limit || 3);
   }
 
   // transform the results and enhance the return values
-  rows = await Promise.all(_.map(rows, async (o) => {
-    const transformedObj = await ctx.mapColumnKeysAndEnhanceOutput(z, bundle, tableMetadata.columns, o);
-    transformedObj.id = `${transformedObj.row_id}-${transformedObj.row_mtime}`;
-    return transformedObj;
-  }));
+  rows = await Promise.all(
+    _.map(rows, async (o) => {
+      const transformedObj = await ctx.mapColumnKeysAndEnhanceOutput(
+        z,
+        bundle,
+        tableMetadata.columns,
+        o
+      );
+      transformedObj.id = `${transformedObj.row_id}-${transformedObj.row_mtime}`;
+      return transformedObj;
+    })
+  );
 
   return rows;
 };
@@ -134,29 +141,29 @@ const perform = async (z, bundle) => {
 const outputFields = async (z, bundle) => {
   const tableMetadata = await ctx.acquireTableMetadata(z, bundle);
   const oF = [
-    {key: "row_id", label: "Original ID"},
-    {key: "row_mtime", label: "Last modification time"},
-    {key: "row_ctime", label: "Creation time"},
+    { key: 'row_id', label: 'Original ID' },
+    { key: 'row_mtime', label: 'Last modification time' },
+    { key: 'row_ctime', label: 'Creation time' },
     ...ctx.outputFieldsRows(tableMetadata.columns, bundle),
   ];
   return oF;
 };
 
 module.exports = {
-  key: "row_update",
-  noun: "Row Update",
+  key: 'row_update',
+  noun: 'Row Update',
   display: {
-    label: "New or Updated Row",
-    description: "Triggers everytime a new or updated row is found.",
+    label: 'New or Updated Row',
+    description: 'Triggers everytime a new or updated row is found.',
   },
   operation: {
     perform,
     inputFields: [ctx.tableFields, ctx.fileNoAuthLinksField],
     sample: {
-      "id": "N33qMZ-JQTuUlx_DiF__lQ",
-      "row_id": "N33qMZ-JQTuUlx_DiF__lQ",
-      "row_mtime": "2021-12-02T01:23:45.678+00:00",
-      "column:0000": "Contents of the first field; a text-field",
+      id: 'N33qMZ-JQTuUlx_DiF__lQ',
+      row_id: 'N33qMZ-JQTuUlx_DiF__lQ',
+      row_mtime: '2021-12-02T01:23:45.678+00:00',
+      'column:0000': 'Contents of the first field; a text-field',
     },
     outputFields: [outputFields],
   },