+++ CHANGELOG.md
@@ -1,5 +1,30 @@
 # Changelog
 
+## v4.0.5 (2025-01-21)
+
+- finalization of "Find Row"
+- finalization of "Find Many Rows
+
+## v4.0.4 (2025-01-20)
+
+- finalization of "API request (Beta)"
+
+## v4.0.3 (2025-01-18)
+
+_skipped_
+
+## v4.0.2 (2025-01-17)
+
+- Updated authentication label
+- finalization of "new asset"
+- finalization of "delete row"
+
+## v4.0.1 (2025-01-17)
+
+- upgrade to zapier-platform-core 16.1.1
+- finalization of "new row"
+- finalization of "new or updated row"
+
 ## v4.0.0 (2024-12-29)
 
 Two triggers read.